refactore - functions are divided into files
